Move Watch to util package
