+++ examples/src/main/java/io/kubernetes/client/examples/WatchExample.java
@@ -14,12 +14,12 @@ package io.kubernetes.client.examples;
 
 import com.google.gson.reflect.TypeToken;
 import io.kubernetes.client.ApiClient;
-import io.kubernetes.client.Watch;
 import io.kubernetes.client.ApiException;
 import io.kubernetes.client.Configuration;
 import io.kubernetes.client.apis.CoreV1Api;
 import io.kubernetes.client.models.V1Namespace;
 import io.kubernetes.client.util.Config;
+import io.kubernetes.client.util.Watch;
 
 import java.io.IOException;
 
@@ -33,7 +33,8 @@ public class WatchExample {
 
         CoreV1Api api = new CoreV1Api();
 
-        Watch<V1Namespace> watch = client.watch(
+        Watch<V1Namespace> watch = Watch.createWatch(
+                client,
                 api.listNamespaceCall(null, null, null, null, 5, Boolean.TRUE, null, null),
                 new TypeToken<Watch.Response<V1Namespace>>(){}.getType());
